update factoring for REST
