+++ packages/json-api/src/request.ts
@@ -1,1 +1,2 @@
 export { findRecord } from './-private/builders/find-record';
+export { query } from './-private/builders/query';
